ICU-5445 Workaround for double precision problem introduced by the change set#21566.
X-SVN-Rev: 22536

+++ icu4j/src/com/ibm/icu/dev/test/calendar/AstroTest.java
@@ -81,7 +81,7 @@ public class AstroTest extends TestFmwk {
                         ") !=  tests[i][5](" + tests[i][5] +
                         ") for test " + i);
             }
-            if (result.declination != tests[i][6]) {
+            if ((float)result.declination != (float)tests[i][6]) {
                 errln("FAIL: result.declination(" + result.declination +
                         ") !=  tests[i][6](" + tests[i][6] +
                         ") for test " + i);
